changed double in float

+++ prod_index.c
@@ -9,7 +9,7 @@
 #define  DEBUG 0
 
 
-void read(double hist[4][4][4]){
+void read(float hist[4][4][4]){
   for (int r = 0; r < 4; ++r){
     for (int g = 0; g < 4; ++g){
       for (int b = 0; b < 4; ++b){
@@ -30,7 +30,7 @@ int main(int argc, char *argv[]){
   char *image = malloc(15*sizeof(char));
   FILE *list = fopen("list.txt", "r");
   FILE *index = fopen("index.txt", "w");
-  double hist[4][4][4];
+  float hist[4][4][4];
   if (list != NULL){
   if (DEBUG) printf("Breakpoint1\n");
   
@@ -52,7 +52,7 @@ int main(int argc, char *argv[]){
     read_cimage(img, &cim);
     for (i = 0; i < 8; i++){
       for (j = 0; j < 8; j++){
-        hist[cim.r[i][j]/64][cim.g[i][j]/64][cim.b[i][j]/64] += (double) 0.015625;
+        hist[cim.r[i][j]/64][cim.g[i][j]/64][cim.b[i][j]/64] += (float) 0.015625;
       }
     }
